+++ src/cron/service/ops.ts
@@ -15,12 +15,10 @@ import {
   markCronJobActive,
   type CronActiveJobMarker,
 } from "../active-jobs.js";
-import { resolveCronDeliveryPlan, resolveFailureDestination } from "../delivery-plan.js";
 import { createCronRunDiagnosticsFromError } from "../run-diagnostics.js";
-import type { CronRunLogEntry } from "../run-log-types.js";
 import { normalizeCronRunLogJobId } from "../run-log.js";
 import { cronSchedulingInputsEqual } from "../schedule-identity.js";
-import type { CronJob, CronJobCreate, CronJobPatch, CronPayload, CronRunStatus } from "../types.js";
+import type { CronJob, CronJobCreate, CronJobPatch, CronPayload } from "../types.js";
 import { normalizeCronRunErrorText } from "./execution-errors.js";
 import { failureNotificationDeliveryFromJobState } from "./failure-alerts.js";
 import {
@@ -47,6 +45,13 @@ import type {
 } from "./list-page-types.js";
 import { locked } from "./locked.js";
 import { normalizeOptionalAgentId } from "./normalize.js";
+import {
+  type InterruptedStartupRun,
+  markInterruptedStartupRun,
+  mergeManualRunSnapshotAfterReload,
+  restoreFinalizedStartupRun,
+  STARTUP_INTERRUPTED_ERROR,
+} from "./startup-run-repair.js";
 import type {
   CronAddOptions,
   CronEvent,
@@ -83,15 +88,6 @@ import {
   stopTimer,
 } from "./timer.js";
 import { wake } from "./wake.js";
-
-const STARTUP_INTERRUPTED_ERROR = "cron: job interrupted by gateway restart";
-
-type InterruptedStartupRun = {
-  jobId: string;
-  taskRunId?: string;
-  runAtMs: number;
-  durationMs: number;
-};
 
 function markManualCronJobActive(
   state: CronServiceState,
@@ -130,150 +126,6 @@ function maybeNotifyManualIsolatedSetupTimeout(
   const notified = maybeNotifyIsolatedAgentSetupTimeout(state, result);
   state.manualSetupTimeoutNotified ||= notified;
   return notified;
-}
-
-function resolveInterruptedStartupFailureNotificationStatus(params: {
-  state: CronServiceState;
-  job: CronJob;
-}) {
-  if (params.job.delivery?.bestEffort === true) {
-    return "not-requested";
-  }
-  if (resolveFailureDestination(params.job, params.state.deps.cronConfig?.failureDestination)) {
-    return "unknown";
-  }
-  const primaryPlan = resolveCronDeliveryPlan(params.job);
-  return primaryPlan.mode === "announce" && primaryPlan.requested ? "unknown" : "not-requested";
-}
-
-function markInterruptedStartupRun(params: {
-  state: CronServiceState;
-  job: CronJob;
-  taskRunId?: string;
-  runningAtMs: number;
-  nowMs: number;
-}): InterruptedStartupRun {
-  const { job, runningAtMs, nowMs } = params;
-  // A persisted running marker means the gateway stopped mid-run; mark it as a
-  // normal failed run so retries, alerts, and run logs all see one outcome.
-  const failureNotificationStatus = resolveInterruptedStartupFailureNotificationStatus({
-    state: params.state,
-    job,
-  });
-  const previousErrors =
-    typeof job.state.consecutiveErrors === "number" && Number.isFinite(job.state.consecutiveErrors)
-      ? Math.max(0, Math.floor(job.state.consecutiveErrors))
-      : 0;
-
-  params.state.deps.log.warn(
-    { jobId: job.id, runningAtMs },
-    "cron: marking interrupted running job failed on startup",
-  );
-
-  job.state.runningAtMs = undefined;
-  job.state.lastRunAtMs = runningAtMs;
-  job.state.lastRunStatus = "error";
-  job.state.lastStatus = "error";
-  job.state.lastError = STARTUP_INTERRUPTED_ERROR;
-  job.state.lastDurationMs = Math.max(0, nowMs - runningAtMs);
-  job.state.consecutiveErrors = previousErrors + 1;
-  job.state.lastDelivered = false;
-  job.state.lastDeliveryStatus = "unknown";
-  job.state.lastDeliveryError = STARTUP_INTERRUPTED_ERROR;
-  job.state.lastFailureNotificationDelivered = undefined;
-  job.state.lastFailureNotificationDeliveryStatus = failureNotificationStatus;
-  job.state.lastFailureNotificationDeliveryError = undefined;
-  job.state.nextRunAtMs = undefined;
-  job.updatedAtMs = nowMs;
-
-  if (job.schedule.kind === "at") {
-    job.enabled = false;
-  }
-
-  return {
-    jobId: job.id,
-    ...(params.taskRunId ? { taskRunId: params.taskRunId } : {}),
-    runAtMs: runningAtMs,
-    durationMs: job.state.lastDurationMs,
-  };
-}
-
-function restoreFinalizedStartupRun(params: {
-  state: CronServiceState;
-  job: CronJob;
-  runningAtMs: number;
-  entry: CronRunLogEntry & { status: CronRunStatus };
-  triggerEval?: CronTriggerEvalOutcome;
-}): boolean {
-  const { state, job, runningAtMs, entry } = params;
-  const startedAt = entry.runAtMs ?? runningAtMs;
-  const shouldDelete = applyJobResult(
-    state,
-    job,
-    {
-      status: entry.status,
-      error: entry.error,
-      deliveryError: entry.deliveryError,
-      diagnostics: entry.diagnostics,
-      delivered: entry.delivered,
-      provider: entry.provider,
-      startedAt,
-      endedAt: entry.ts,
-    },
-    { replayFailureAlertAtMs: entry.ts },
-  );
-
-  // The finalized row captured post-run state before the stale cron store write.
-  job.state.lastDurationMs = entry.durationMs ?? Math.max(0, entry.ts - startedAt);
-  job.state.lastErrorReason = entry.errorReason;
-  job.state.lastDelivered = entry.delivered;
-  job.state.lastDeliveryStatus = entry.deliveryStatus;
-  job.state.lastDeliveryError = entry.deliveryError;
-  job.state.lastFailureNotificationDelivered = entry.failureNotificationDelivery?.delivered;
-  job.state.lastFailureNotificationDeliveryStatus = entry.failureNotificationDelivery?.status;
-  job.state.lastFailureNotificationDeliveryError = entry.failureNotificationDelivery?.error;
-  job.state.nextRunAtMs = entry.nextRunAtMs;
-  if (params.triggerEval) {
-    applyTriggerRunResult(job, {
-      status: entry.status,
-      endedAt: entry.ts,
-      triggerEval: params.triggerEval,
-    });
-  }
-  state.deps.log.info(
-    { jobId: job.id, runningAtMs, status: entry.status },
-    "cron: restored finalized task-ledger run on startup",
-  );
-  return shouldDelete;
-}
-
-function mergeManualRunSnapshotAfterReload(params: {
-  state: CronServiceState;
-  jobId: string;
-  snapshot: {
-    enabled: boolean;
-    updatedAtMs: number;
-    state: CronJob["state"];
-  } | null;
-  removed: boolean;
-}) {
-  if (!params.state.store) {
-    return;
-  }
-  if (params.removed) {
-    params.state.store.jobs = params.state.store.jobs.filter((job) => job.id !== params.jobId);
-    return;
-  }
-  if (!params.snapshot) {
-    return;
-  }
-  const reloaded = params.state.store.jobs.find((job) => job.id === params.jobId);
-  if (!reloaded) {
-    return;
-  }
-  reloaded.enabled = params.snapshot.enabled;
-  reloaded.updatedAtMs = params.snapshot.updatedAtMs;
-  reloaded.state = params.snapshot.state;
 }
 
 async function ensureLoadedForRead(state: CronServiceState) {
