+++ src/state/openclaw-state-db-schema-helpers.ts
@@ -1,0 +1,35 @@
+// Provides shared SQLite schema probes and additive column migration helpers.
+import type { DatabaseSync } from "node:sqlite";
+
+export function tableHasColumn(db: DatabaseSync, tableName: string, columnName: string): boolean {
+  const rows = db.prepare(`PRAGMA table_info(${tableName})`).all() as Array<{ name?: unknown }>;
+  return rows.some((row) => row.name === columnName);
+}
+
+export function tablePrimaryKeyColumns(db: DatabaseSync, tableName: string): string[] {
+  const rows = db.prepare(`PRAGMA table_info(${tableName})`).all() as Array<{
+    name?: unknown;
+    pk?: unknown;
+  }>;
+  return rows
+    .filter((row) => Number(row.pk ?? 0) > 0 && typeof row.name === "string")
+    .toSorted((left, right) => Number(left.pk ?? 0) - Number(right.pk ?? 0))
+    .map((row) => row.name as string);
+}
+
+export function tableExists(db: DatabaseSync, tableName: string): boolean {
+  const row = db
+    .prepare("SELECT 1 AS ok FROM sqlite_master WHERE type = 'table' AND name = ?")
+    .get(tableName) as { ok?: unknown } | undefined;
+  return row?.ok === 1;
+}
+
+export function ensureColumn(db: DatabaseSync, tableName: string, columnSql: string): boolean {
+  const columnName = columnSql.trim().split(/\s+/, 1)[0];
+  if (!columnName || !tableExists(db, tableName) || tableHasColumn(db, tableName, columnName)) {
+    return false;
+  }
+  // State migrations are additive here; destructive or shape-changing repairs belong in doctor.
+  db.exec(`ALTER TABLE ${tableName} ADD COLUMN ${columnSql};`);
+  return true;
+}
