+++ src/tasks/task-registry.ts
@@ -47,6 +47,11 @@ import {
   updateFlowRecordByIdExpectedRevision,
 } from "./task-flow-runtime-internal.js";
 import type { TaskRegistryControlRuntime } from "./task-registry-control.types.js";
+import {
+  cloneTaskDeliveryState,
+  cloneTaskRecord,
+  normalizeTaskTimestamps,
+} from "./task-registry-records.js";
 import { getTaskRegistryProcessState } from "./task-registry.process-state.js";
 import {
   getTaskRegistryObservers,
@@ -238,66 +243,6 @@ function ensureTaskCancellationReady(task: TaskRecord): void {
   for (const linkedTask of linkedTasks.length > 0 ? linkedTasks : [task]) {
     ensureLinkedTaskFlowRegistryReady(linkedTask);
   }
-}
-
-function cloneTaskRecord(record: TaskRecord): TaskRecord {
-  return {
-    ...record,
-    ...(record.detail !== undefined ? { detail: structuredClone(record.detail) } : {}),
-  };
-}
-
-function normalizeTaskTimestamps(task: TaskRecord): TaskRecord {
-  // Detached runtimes can report lifecycle times captured before the registry
-  // inserted or restored the row; keep createdAt as the visible lifecycle floor.
-  let createdAt = task.createdAt;
-  for (const candidate of [task.startedAt, task.lastEventAt, task.endedAt]) {
-    if (typeof candidate === "number" && candidate < createdAt) {
-      createdAt = candidate;
-    }
-  }
-
-  const startedAt =
-    typeof task.startedAt === "number" ? Math.max(task.startedAt, createdAt) : task.startedAt;
-  const lastEventAt =
-    typeof task.lastEventAt === "number"
-      ? Math.max(task.lastEventAt, startedAt ?? createdAt)
-      : task.lastEventAt;
-  const endedAt =
-    typeof task.endedAt === "number"
-      ? Math.max(task.endedAt, startedAt ?? createdAt)
-      : task.endedAt;
-
-  if (
-    createdAt === task.createdAt &&
-    startedAt === task.startedAt &&
-    lastEventAt === task.lastEventAt &&
-    endedAt === task.endedAt
-  ) {
-    return task;
-  }
-
-  const normalized: TaskRecord = {
-    ...task,
-    createdAt,
-  };
-  if (typeof startedAt === "number") {
-    normalized.startedAt = startedAt;
-  }
-  if (typeof lastEventAt === "number") {
-    normalized.lastEventAt = lastEventAt;
-  }
-  if (typeof endedAt === "number") {
-    normalized.endedAt = endedAt;
-  }
-  return normalized;
-}
-
-function cloneTaskDeliveryState(state: TaskDeliveryState): TaskDeliveryState {
-  return {
-    ...state,
-    ...(state.requesterOrigin ? { requesterOrigin: { ...state.requesterOrigin } } : {}),
-  };
 }
 
 function snapshotTaskRecords(source: ReadonlyMap<string, TaskRecord>): TaskRecord[] {
